requirements.txt
python-telegram-bot==3.4
sqlalchemy==1.0.12
python-memcached==1.57